Refactors the code to make it easier to maintain
The repositories is now holding the seeds as items and the code generation is done within the seed

+++ Distancify.Migrations.Litium/SeedBuilder/LitiumGraphqlModel/Country.cs
@@ -9,6 +9,7 @@ namespace Distancify.Migrations.Litium.SeedBuilder.LitiumGraphqlModel
         public Currency Currency { get; set; }
 
         //taxclass
-        public double? StandardVatRate { get; set; }
+        public decimal? StandardVatRate { get; set; }
+
     }
 }
